refactor:画像生成関数をapiの中に移動
